Feat: выполнен перевод обратно на русский язык
Fix: перевод на английский добавлен

Также удалил папку asserts, что немного отчистило её
