+++ Сайт/Export_script/bootstrap_export_script.py
@@ -4,6 +4,8 @@
 import sys
 from os import path, listdir
 from bs4 import BeautifulSoup
+from bs4.element import Comment
+from re import match as re_match
 
 main_path = sys.argv[1]
 template_path = path.join(main_path, "templates")
@@ -13,6 +15,17 @@ def getDirsList(path):
 
 def getFilesList(directory):
     return [path.join(directory, fileName) for fileName in getDirsList(directory) if path.isfile(path.join(directory, fileName))]
+
+def is_visible(element):
+    if element in [' ', '\n']:
+        return False
+    if element.parent.name in ['style', 'script', 'head', 'title', 'meta', '[document]']:
+        return False
+    if isinstance(element, Comment):
+        return False
+    if re_match(r"[^>]*{{[^{}]*}}[^<]*", element) != None or re_match(r"[^>]*{%[^{}%]+%}[^<]*", element) != None:
+        return False
+    return True
 
 for template in getFilesList(template_path):
     file_data = str()
@@ -50,5 +63,10 @@ for template in getFilesList(template_path):
         first_script.insert_before("{% block scripts %}")
         last_script.insert_after("{% endblock scripts %}")
 
-    with open('test', 'wb') as file:
+    texts = soup.findAll(text=True)
+    visible_texts = filter(is_visible, texts)
+    for visible_text in visible_texts:
+        visible_text.replace_with("{{ _(\"" + str(visible_text) + "\") }}")
+
+    with open(template, 'wb') as file:
         file.write(soup.prettify("utf-8"))
