+++ Сайт/Website/templates/project.html
@@ -1,42 +1,56 @@
 <!DOCTYPE html>
 <html>
-
-<head>
-    <meta charset="utf-8">
-{% block head %}
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, shrink-to-fit=no">
-    <title>RidrameCraft</title>
-    <meta name="description" content="Projects page">
-    <meta property="og:type" content="website">
-    <link rel="icon" type="image/png" sizes="384x384" href="../assets/img/website_icon.png?h=15b8e40d28339fb582ea1179e6313669">
-    <link rel="icon" type="image/png" sizes="384x384" href="../assets/img/website_icon.png?h=15b8e40d28339fb582ea1179e6313669">
-    <link rel="stylesheet" href="../assets/bootstrap/css/bootstrap.min.css?h=cb9114bd32b0a029d9eedd9db9c0b8e8">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Montserrat:400,700">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Noto+Sans:400,700&amp;subset=cyrillic&amp;display=swap">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Open+Sans:400,700&amp;subset=cyrillic&amp;display=swap">
-    <link rel="stylesheet" href="../assets/fonts/font-awesome.min.css?h=ef9efa29ed5490562c2ff272112d2e97">
-    <link rel="stylesheet" href="../assets/css/styles.min.css?h=9955dc8c698741fcb5ccc6399f603354">
-    <link rel="stylesheet" href="../assets/css/Projects-Horizontal.min.css?h=9a1bf293ff01178d21bc6b4bc73bc202">
-    <link rel="stylesheet" href="../assets/css/projects.min.css?h=ec3f70e9591058f13632ba797c20e6ba">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css">
-<script src="{{ url_for('babel_catalog') }}"></script>
-{% endblock head %}
-</head>
-
-<body id="pop-up_window" style="background: rgb(44,62,80);border-radius: 0px;padding: 0px;">
-    <div class="d-flex flex-column" id="window_message"><a class="d-flex flex-shrink-0 justify-content-center align-items-center align-self-end" style="background: #18bc9c;margin: 8px;border-radius: 100%;border-style: none;color: var(--light);width: 40px;height: 40px;text-decoration: none; cursor: pointer;" onclick="closeProject();">X</a>
-        <div></div>
-        <div class="d-flex flex-column flex-fill justify-content-center align-items-center justify-content-xl-center align-items-xl-center">
-            <h1 class="main_text">{{ _("Project name: ") + project_name}}</h1>
-        </div>
-    </div>
-    <script src="../assets/js/jquery.min.js?h=84e399b8f2181ccd73394fdeddff1638"></script>
-    <script src="../assets/bootstrap/js/bootstrap.min.js?h=2504f2315ca47ea4d62e67e20a5551d7"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js"></script>
-    <script src="https://cdn.datatables.net/1.10.20/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.10.20/js/dataTables.bootstrap4.min.js"></script>
-    <script src="../assets/js/script.min.js?h=90ac2f229acb6b4e5f56600e1c50bada"></script>
-</body>
-
+ <head>
+  <meta charset="utf-8"/>
+  {% block head %}
+  <meta content="width=device-width, initial-scale=1.0, shrink-to-fit=no" name="viewport"/>
+  <title>
+   RidrameCraft
+  </title>
+  <meta content="Projects page" name="description"/>
+  <meta content="website" property="og:type"/>
+  <link href="../assets/img/website_icon.png?h=15b8e40d28339fb582ea1179e6313669" rel="icon" sizes="384x384" type="image/png"/>
+  <link href="../assets/img/website_icon.png?h=15b8e40d28339fb582ea1179e6313669" rel="icon" sizes="384x384" type="image/png"/>
+  <link href="../assets/bootstrap/css/bootstrap.min.css?h=a5f1ef960f6df65e5239617e2e8cfb70" rel="stylesheet"/>
+  <link href="https://fonts.googleapis.com/css?family=Montserrat:400,700" rel="stylesheet"/>
+  <link href="https://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic" rel="stylesheet"/>
+  <link href="https://fonts.googleapis.com/css?family=Noto+Sans:400,700&amp;subset=cyrillic&amp;display=swap" rel="stylesheet"/>
+  <link href="https://fonts.googleapis.com/css?family=Open+Sans:400,700&amp;subset=cyrillic&amp;display=swap" rel="stylesheet"/>
+  <link href="../assets/fonts/font-awesome.min.css?h=ef9efa29ed5490562c2ff272112d2e97" rel="stylesheet"/>
+  <link href="../assets/css/styles.min.css?h=9955dc8c698741fcb5ccc6399f603354" rel="stylesheet"/>
+  <link href="../assets/css/Projects-Horizontal.min.css?h=9a1bf293ff01178d21bc6b4bc73bc202" rel="stylesheet"/>
+  <link href="../assets/css/projects.min.css?h=ec3f70e9591058f13632ba797c20e6ba" rel="stylesheet"/>
+  <link href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.5.2/animate.min.css" rel="stylesheet"/>
+  <script src="{{ url_for('babel_catalog') }}">
+  </script>
+  {% endblock head %}
+ </head>
+ <body id="pop-up_window" style="background: rgb(44,62,80);border-radius: 0px;padding: 0px;">
+  <div class="d-flex flex-column" id="window_message">
+   <a class="d-flex flex-shrink-0 justify-content-center align-items-center align-self-end" onclick="closeProject();" style="background: #18bc9c;margin: 8px;border-radius: 100%;border-style: none;color: var(--light);width: 40px;height: 40px;text-decoration: none; cursor: pointer;">
+    {{ _("X") }}
+   </a>
+   <div>
+   </div>
+   <div class="d-flex flex-column flex-fill justify-content-center align-items-center justify-content-xl-center align-items-xl-center">
+    <h1 class="main_text">
+     {{ _("Название проекта: ") + project_name}}
+    </h1>
+   </div>
+  </div>
+  {% block scripts %}
+  <script src="../assets/js/jquery.min.js?h=84e399b8f2181ccd73394fdeddff1638">
+  </script>
+  <script src="../assets/bootstrap/js/bootstrap.min.js?h=2504f2315ca47ea4d62e67e20a5551d7">
+  </script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-easing/1.4.1/jquery.easing.min.js">
+  </script>
+  <script src="https://cdn.datatables.net/1.10.20/js/jquery.dataTables.min.js">
+  </script>
+  <script src="https://cdn.datatables.net/1.10.20/js/dataTables.bootstrap4.min.js">
+  </script>
+  <script src="../assets/js/script.min.js?h=90ac2f229acb6b4e5f56600e1c50bada">
+  </script>
+  {% endblock scripts %}
+ </body>
 </html>
